[stm.js] Break a long transport stream from ffmpeg up into chunks, instead of creating the short chunks directly.
This fixes the audio (and sometimes video) glitching on every chunk
transition.

This works by interpretting the mpeg transport stream with mpeg2ts.js, which
watches the PCR timestamp. Chunks are on mpeg packet boundaries.

Any discontinuity causes the ffmpeg process to start over with another long
running transport stream.

As a side effect, none of the data is ever written to the filesystem any
more. It's all piped to us directly and held in buffers.